backend/requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
requests==2.31.0
